Working Overlay Joystick and Button
+ Update button struct to handle line_height, centered text or not, etc..
+ Setup handler for keypress button
+ Setup button color on draw
+ Update CI to install dependencies

+++ src/graphics/utils.rs
@@ -35,5 +35,5 @@ pub fn get_texture_rect(
     } else {
         (x, y)
     };
-    Rect::new(cx as i32, cy as i32, w as u32, h as u32)
+    Rect::new(cx, cy, w as u32, h as u32)
 }
